requirements.txt
asgiref==3.10.0
click==8.3.0
colorama==0.4.6
cssbeautifier==1.15.4
dj-database-url==3.0.1
Django==5.2.7
django-extensions==4.1
django-showurls==1.0.1
djlint==1.36.4
django-anymail==13.1
EditorConfig==0.17.1
fire==0.7.1
fonttools==4.60.1
git-filter-repo==2.47.0
gunicorn==23.0.0
jsbeautifier==1.15.4
json5==0.12.1
lxml==6.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pathspec==0.12.1
pdf2docx==0.5.8
PyMuPDF==1.26.4
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.3
regex==2025.9.18
six==1.17.0
sqlparse==0.5.3
termcolor==3.1.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
waitress==3.0.2
whitenoise==6.11.0
